chore: merge synch and sync package
as synch was likely just a typo
