+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/dxf2/sync/SyncUtilsTest.java
@@ -30,7 +30,6 @@ package org.hisp.dhis.dxf2.sync;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
-import org.hisp.dhis.dxf2.synch.SystemInstance;
 import org.hisp.dhis.setting.SettingKey;
 import org.hisp.dhis.setting.SystemSettingManager;
 import org.hisp.dhis.test.integration.SingleSetupIntegrationTestBase;
